requirements.txt
# Whisper-Input v2.0.0 Dependencies
# Enhanced Voice Transcription Tool

# Core dependencies
aiohttp>=3.9.0
annotated-types==0.7.0
anyio==4.8.0
build==1.2.2.post1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
distro==1.9.0
exceptiongroup==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.1
openai==1.59.6
opencc-python-reimplemented==0.1.7
packaging==24.2
pip-tools==7.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-core==2.27.2
pynput==1.7.7
pyobjc-core==10.3.2
pyobjc-framework-applicationservices==10.3.2
pyobjc-framework-cocoa==10.3.2
pyobjc-framework-coretext==10.3.2
pyobjc-framework-quartz==10.3.2
pyperclip==1.9.0
pyproject-hooks==1.2.0
pyqt5==5.15.11
pyqt5-qt5==5.15.16
pyqt5-sip==12.16.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1